refactoring views and serializers
