+++ api/v1/serializers/merch_serializer.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+
+from ambassadors.models import Merch
+
+
+class MerchSerializer(serializers.ModelSerializer):
+    """Сериализатор мечра."""
+
+    class Meta:
+        model = Merch
+        fields = '__all__'
