+++ api/v1/views/__init__.py
@@ -1,0 +1,10 @@
+from .address_view import AddressViewSet
+from .ambassador_view import AmbassadorsViewSet
+from .content_view import ContentViewSet
+from .merch_view import MerchViewSet
+from .notification_view import notification_list, notification_detail
+from .profile_view import ProfileViewSet
+from .promocode_view import PromocodeViewSet
+from .sent_merch_view import SentMerchViewSet
+from .user_view import UserAPIView
+from .ya_form_view import ambassadors_form_get
