+++ api/v1/views/address_view.py
@@ -1,0 +1,12 @@
+from rest_framework import viewsets
+from rest_framework.permissions import IsAuthenticated
+
+from ambassadors.models import Address
+from api.v1.serializers import AddressSerializer
+
+
+class AddressViewSet(viewsets.ModelViewSet):
+    """Вьюсет для адреса."""
+    permission_classes = [IsAuthenticated, ]
+    serializer_class = AddressSerializer
+    queryset = Address.objects.all()
