+++ api/v1/views/ambassador_view.py
@@ -1,0 +1,38 @@
+from http import HTTPMethod
+
+from django.utils import timezone
+from drf_spectacular.utils import extend_schema
+from rest_framework import viewsets, mixins
+from rest_framework.permissions import IsAuthenticated
+
+from ambassadors.models import Ambassador
+from api.v1.serializers import (
+    AmbassadorReadSerializer, AmbassadorWriteSerializer
+)
+
+AMBASSADORS_DESCRIPTION = (
+    'Эндпоинты для создания, изменения и просмотра амбассадоров'
+)
+
+
+@extend_schema(tags=['Амбассадоры'], description=AMBASSADORS_DESCRIPTION)
+class AmbassadorsViewSet(
+    viewsets.GenericViewSet,
+    mixins.RetrieveModelMixin,
+    mixins.ListModelMixin,
+    mixins.CreateModelMixin,
+    mixins.UpdateModelMixin
+):
+    http_method_names = ['get', 'post', 'patch']
+    permission_classes = [IsAuthenticated, ]
+    queryset = Ambassador.objects.all()
+
+    def get_serializer_class(self):
+        match self.request.method:
+            case HTTPMethod.GET:
+                return AmbassadorReadSerializer
+            case HTTPMethod.POST | HTTPMethod.PATCH:
+                return AmbassadorWriteSerializer
+
+    def perform_create(self, serializer):
+        serializer.save(pub_date=timezone.now())
